move to using CUtlVectorMemory over CUtlMemory and add CUtlVectorMemory_Growable
