+++ tier1/tier1.vcxproj
@@ -368,6 +368,7 @@
     <ClInclude Include="..\public\tier1\utllinkedlist.h" />
     <ClInclude Include="..\public\tier1\utlmap.h" />
     <ClInclude Include="..\public\tier1\utlmemory.h" />
+    <ClInclude Include="..\public\tier1\utlvectormemory.h" />
     <ClInclude Include="..\public\tier1\utlmultilist.h" />
     <ClInclude Include="..\public\tier1\utlpriorityqueue.h" />
     <ClInclude Include="..\public\tier1\utlqueue.h" />
